+++ webrtc/modules/audio_processing/beamformer/covariance_matrix_generator.h
@@ -11,8 +11,6 @@
 #ifndef WEBRTC_MODULES_AUDIO_PROCESSING_BEAMFORMER_COVARIANCE_MATRIX_GENERATOR_H_
 #define WEBRTC_MODULES_AUDIO_PROCESSING_BEAMFORMER_COVARIANCE_MATRIX_GENERATOR_H_
 
-#include <vector>
-
 #include "webrtc/modules/audio_processing/beamformer/complex_matrix.h"
 #include "webrtc/modules/audio_processing/beamformer/array_util.h"
 
@@ -32,9 +30,9 @@ class CovarianceMatrixGenerator {
   // The covariance matrix of a source at the given angle.
   static void AngledCovarianceMatrix(float sound_speed,
                                      float angle,
-                                     int frequency_bin,
-                                     int fft_size,
-                                     int num_freq_bins,
+                                     size_t frequency_bin,
+                                     size_t fft_size,
+                                     size_t num_freq_bins,
                                      int sample_rate,
                                      const std::vector<Point>& geometry,
                                      ComplexMatrix<float>* mat);
@@ -42,8 +40,8 @@ class CovarianceMatrixGenerator {
   // Calculates phase shifts that, when applied to a multichannel signal and
   // added together, cause constructive interferernce for sources located at
   // the given angle.
-  static void PhaseAlignmentMasks(int frequency_bin,
-                                  int fft_size,
+  static void PhaseAlignmentMasks(size_t frequency_bin,
+                                  size_t fft_size,
                                   int sample_rate,
                                   float sound_speed,
                                   const std::vector<Point>& geometry,
