+++ webrtc/modules/utility/interface/file_player.h
@@ -20,8 +20,9 @@
 namespace webrtc {
 class FileCallback;
 
-class FilePlayer {
- public:
+class FilePlayer
+{
+public:
     // The largest decoded frame size in samples (60ms with 32kHz sample rate).
     enum {MAX_AUDIO_BUFFER_IN_SAMPLES = 60*32};
     enum {MAX_AUDIO_BUFFER_IN_BYTES = MAX_AUDIO_BUFFER_IN_SAMPLES*2};
@@ -37,7 +38,7 @@ class FilePlayer {
     // channel).
     virtual int Get10msAudioFromFile(
         int16_t* outBuffer,
-        int& lengthInSamples,
+        size_t& lengthInSamples,
         int frequencyInHz) = 0;
 
     // Register callback for receiving file playing notifications.
@@ -102,8 +103,9 @@ class FilePlayer {
       return -1;
     }
 
- protected:
+protected:
     virtual ~FilePlayer() {}
+
 };
 }  // namespace webrtc
-#endif  // WEBRTC_MODULES_UTILITY_INTERFACE_FILE_PLAYER_H_
+#endif // WEBRTC_MODULES_UTILITY_INTERFACE_FILE_PLAYER_H_
